Fix: fix searching and loading question logic part of SWM-154 related to #59
